create memberShipController and move invite and accept in this controller
